MIG-63: Moved script inside MenuConfigMapping
