+++ packages/mdctl-axon-tools/lib/StudyManifestTools.js
@@ -135,11 +135,10 @@ class StudyManifestTools {
           { outputEntities, removedEntities } = this.validateReferences(allEntities, orgReferenceProps),
           manifest = this.createManifest(outputEntities),
           menuConfigMapping = new MenuConfigMapping(org),
-          mappings = await menuConfigMapping.getMappings(),
-          hasMappings = !!mappings.length
+          mappingScript = await menuConfigMapping.getMappingScript()
 
-    if (hasMappings) {
-      this.writePostImportScript(mappings)
+    if (mappingScript) {
+      this.writeInstallAftertScript(mappingScript)
     }
 
     this.writeIssues(removedEntities)
@@ -148,54 +147,13 @@ class StudyManifestTools {
     return { manifest, removedEntities }
   }
 
-  async writePostImportScript(mappings) {
+  writeInstallAftertScript(mappingScript) {
     const { options } = privatesAccessor(this),
           outputDir = options.dir || process.cwd()
 
     console.log('Writing post import script')
 
-    fs.writeFileSync(`${outputDir}/install.after.js`, `
-const { run } = require('expressions')
-
-const mappings = ${JSON.stringify(mappings)}
-
-mappings.forEach(({ path, mapTo }) => {
-  const [entity, entityKey, property, ...rest] = path.split('.'),
-      isDocPropUpdate = !!rest.length,
-      value = run(mapTo)
-
-if (isDocPropUpdate) {
-  const [entityResult] = org.objects[entity]
-    .find({ c_key: entityKey })
-    .paths(property)
-    .limit(1)
-    .toArray()
-
-  if (!entityResult) return
-
-  const documentProps = entityResult[property]
-
-  if (!documentProps || documentProps.length === 0) return
-
-  const [docPropKey, docProp] = rest
-
-  if (!docPropKey || !docProp) return
-
-  const propToUpdate = documentProps.find(({ c_key }) => c_key === docPropKey),
-
-        idToUpdate = propToUpdate._id
-
-  return org.objects[entity]
-    .updateOne({ c_key: entityKey })
-    .pathUpdate(property + '/' + idToUpdate + '/' + docProp , value)
-
-}
-
-return org.objects[entity]
-  .updateOne({ c_key: entityKey }, { $set: { [property]: value } })
-  .execute()
-})
-    `)
+    fs.writeFileSync(`${outputDir}/install.after.js`, mappingScript)
   }
 
   async getTasksManifest(taskIds) {
